RF: organize conftest.py; move Faker provider to `onyo/lib/`
